+++ snowddl/app/__init__.py
@@ -1,2 +1,3 @@
 from .base import BaseApp
 from .convert import ConvertApp
+from .singledb import SingleDbApp