+++ snowddl/app/convert.py
@@ -44,39 +44,39 @@ class ConvertApp(BaseApp):
         return parser
 
     def init_config_path(self):
-        config_path = Path(self.args.c)
+        config_path = Path(self.args['c'])
 
-        if config_path.is_dir() and self.args.clean:
+        if config_path.is_dir() and self.args.get('clean'):
             rmtree(config_path)
 
         if not config_path.exists():
             config_path.mkdir(mode=0o755, parents=True)
 
         if not config_path.is_dir():
-            raise ValueError(f"Config path [{self.args.c}] is not a directory")
+            raise ValueError(f"Config path [{self.args['c']}] is not a directory")
 
         return config_path
 
     def init_config(self):
-        return SnowDDLConfig(self.args.env_prefix)
+        return SnowDDLConfig(self.args.get('env_prefix'))
 
     def init_settings(self):
         settings = SnowDDLSettings()
 
-        if self.args.exclude_object_types:
+        if self.args.get('exclude_object_types'):
             try:
-                settings.exclude_object_types = [ObjectType[t.strip().upper()] for t in str(self.args.exclude_object_types).split(',')]
+                settings.exclude_object_types = [ObjectType[t.strip().upper()] for t in str(self.args.get('exclude_object_types')).split(',')]
             except KeyError as e:
                 raise ValueError(f"Invalid object type [{str(e)}]")
 
-        if self.args.include_object_types:
+        if self.args.get('include_object_types'):
             try:
-                settings.include_object_types = [ObjectType[t.strip().upper()] for t in str(self.args.include_object_types).split(',')]
+                settings.include_object_types = [ObjectType[t.strip().upper()] for t in str(self.args.get('include_object_types')).split(',')]
             except KeyError as e:
                 raise ValueError(f"Invalid object type [{str(e)}]")
 
-        if self.args.max_workers:
-            settings.max_workers = self.args.max_workers
+        if self.args.get('max_workers'):
+            settings.max_workers = int(self.args.get('max_workers'))
 
         return settings