+++ snowddl/resolver/row_access_policy.py
@@ -1,6 +1,6 @@
 from json import loads
 
-from snowddl.blueprint import DatabaseIdent, RowAccessPolicyBlueprint, ObjectType, Edition, SchemaObjectIdent
+from snowddl.blueprint import RowAccessPolicyBlueprint, ObjectType, Edition, SchemaObjectIdent
 from snowddl.resolver.abc_schema_object_resolver import AbstractSchemaObjectResolver, ResolveResult
 
 
@@ -183,7 +183,7 @@ class RowAccessPolicyResolver(AbstractSchemaObjectResolver):
         existing_policy_refs = {}
 
         cur = self.engine.execute_meta("SELECT * FROM TABLE({database:i}.information_schema.policy_references(policy_name => {policy_name}))", {
-            "database": DatabaseIdent(policy_name.env_prefix, policy_name.database),
+            "database": policy_name.database_full_name,
             "policy_name": policy_name,
         })